# docs/openapi.yaml
openapi: 3.0.3
info:
  title: PageFire API
  version: 0.3.0
  description: |
    PageFire is a unified monitoring and incident management platform.
    Single Go binary, self-hosted, MIT license.

    Authentication is via cookie-based sessions (login endpoint) or
    Bearer API tokens (created via the tokens endpoint). Admin-only
    write operations are enforced on users, teams, services, escalation
    policies, and schedules.
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  contact:
    url: https://pagefire.dev
servers:
  - url: /api/v1
    description: Default API base path
...
security:
  - cookieAuth: []
  - bearerAuth: []
paths:
...
  /users:
    post:
      operationId: createUser
      tags: [Users]
      summary: Create a new user (admin only)
      description: |
        If no password is provided, an invite token is generated and returned as invite_url.
        Only admins can set role to "admin".
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateUserRequest"
      responses:
        "201":
          description: User created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CreateUserResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
    get:
      operationId: listUsers
      tags: [Users]
      summary: List all users
      responses:
        "200":
          description: User list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/User"
...
  /teams:
    post:
      operationId: createTeam
      tags: [Teams]
      summary: Create a new team (admin only)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateTeamRequest"
      responses:
        "201":
          description: Team created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Team"
        "400":
          $ref: "#/components/responses/BadRequest"
    get:
      operationId: listTeams
      tags: [Teams]
      summary: List all teams
      responses:
        "200":
          description: Team list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Team"
...
  /services:
    post:
      operationId: createService
      tags: [Services]
      summary: Create a new service (admin only)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateServiceRequest"
      responses:
        "201":
          description: Service created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Service"
        "400":
          $ref: "#/components/responses/BadRequest"
    get:
      operationId: listServices
      tags: [Services]
      summary: List all services
      parameters:
        - name: team_id
          in: query
          description: Filter services by team ID
          schema:
            type: string
      responses:
        "200":
          description: Service list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Service"
...
  /escalation-policies:
    post:
      operationId: createEscalationPolicy
      tags: [Escalation Policies]
      summary: Create an escalation policy (admin only)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateEscalationPolicyRequest"
      responses:
        "201":
          description: Escalation policy created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EscalationPolicy"
        "400":
          $ref: "#/components/responses/BadRequest"
    get:
      operationId: listEscalationPolicies
      tags: [Escalation Policies]
      summary: List all escalation policies
      parameters:
        - name: team_id
          in: query
          description: Filter by team ID
          schema:
            type: string
      responses:
        "200":
          description: Escalation policy list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/EscalationPolicy"
...
  /schedules:
    post:
      operationId: createSchedule
      tags: [Schedules]
      summary: Create a schedule (admin only)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateScheduleRequest"
      responses:
        "201":
          description: Schedule created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Schedule"
        "400":
          $ref: "#/components/responses/BadRequest"
    get:
      operationId: listSchedules
      tags: [Schedules]
      summary: List all schedules
      parameters:
        - name: team_id
          in: query
          description: Filter by team ID
          schema:
            type: string
      responses:
        "200":
          description: Schedule list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Schedule"
...
  /alerts:
    post:
      operationId: createAlert
      tags: [Alerts]
      summary: Create a new alert
      description: Triggers the escalation pipeline for the specified service.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateAlertRequest"
      responses:
        "201":
          description: Alert created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Alert"
        "400":
          $ref: "#/components/responses/BadRequest"
        "404":
          description: Service not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
    get:
      operationId: listAlerts
      tags: [Alerts]
      summary: List alerts with filtering and pagination
      parameters:
        - name: status
          in: query
          description: "Filter by status: triggered, acknowledged, resolved"
          schema:
            type: string
            enum: [triggered, acknowledged, resolved]
        - name: service_id
          in: query
          description: Filter by service ID
          schema:
            type: string
        - name: group_key
          in: query
          description: Filter by group key
          schema:
            type: string
        - name: source
          in: query
          description: "Filter by source: api, integration, monitor, test"
          schema:
            type: string
        - name: search
          in: query
          description: Full-text search in summary (max 200 chars)
          schema:
            type: string
            maxLength: 200
        - name: created_after
          in: query
          description: Filter alerts created after this time (RFC3339)
          schema:
            type: string
            format: date-time
        - name: created_before
          in: query
          description: Filter alerts created before this time (RFC3339)
          schema:
            type: string
            format: date-time
        - name: limit
          in: query
          description: Maximum number of results (default 50, max 200)
          schema:
            type: integer
            default: 50
            maximum: 200
        - name: offset
          in: query
          description: Offset for pagination
          schema:
            type: integer
            default: 0
      responses:
        "200":
          description: Alert list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Alert"
...
  /incidents:
    post:
      operationId: createIncident
      tags: [Incidents]
      summary: Create a new incident
      description: Defaults to status "triggered", severity "critical", source "manual".
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateIncidentRequest"
      responses:
        "201":
          description: Incident created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Incident"
        "400":
          $ref: "#/components/responses/BadRequest"
    get:
      operationId: listIncidents
      tags: [Incidents]
      summary: List incidents with filtering and pagination
      parameters:
        - name: status
          in: query
          description: "Filter by status"
          schema:
            type: string
            enum: [triggered, acknowledged, investigating, identified, monitoring, resolved]
        - name: search
          in: query
          description: Full-text search in title (max 200 chars)
          schema:
            type: string
            maxLength: 200
        - name: limit
          in: query
          schema:
            type: integer
            default: 50
            maximum: 200
        - name: offset
          in: query
          schema:
            type: integer
            default: 0
      responses:
        "200":
          description: Incident list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Incident"
...
  /events:
    get:
      operationId: listHomeEvents
      tags: [Home Alerts]
      summary: List the current user's 20 most recent home events
      responses:
        '200': { description: Recent event records, including playground events }
    post:
      operationId: createHomeEvent
      tags: [Home Alerts]
      summary: Submit a start, stop, or info event
      description: The client supplies event_id. Duplicate IDs and active start events do not send another notification. Accepts a session, personal API token, or the single event-only ingestion key from Event Ingestion settings.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [event_id, event]
              properties:
                event_id: { type: string, maxLength: 128 }
                event: { type: string, enum: [start, stop, info] }
                incident_key: { type: string, description: Required for start and stop }
                severity: { type: string, enum: [low, mid, high], description: Required for start; info defaults to low }
                summary: { type: string, maxLength: 500, description: Required for start and info }
                details: { type: string, maxLength: 10000 }
                repeat_interval_seconds: { type: integer, minimum: 300, default: 300, description: High start events only }
      responses:
        '201': { description: Event applied }
        '200': { description: Duplicate or inactive event ignored }
        '400': { description: Invalid event }
  /events/{event_id}:
    get:
      operationId: getHomeEvent
      tags: [Home Alerts]
      summary: Inspect an event and its plugin delivery attempts
      parameters:
        - in: path
          name: event_id
          required: true
          schema: { type: string }
      responses:
        '200': { description: Event and delivery status }
        '404': { description: Event not found }
  /alert-events:
    get:
      operationId: listCombinedAlertEvents
      tags: [Home Alerts]
      summary: Filter the combined Home Events and service-alert feed
      parameters:
        - in: query
          name: window
          schema: { type: string, enum: [1h, 1d, 1w, 1m], default: 1d }
        - in: query
          name: type
          schema: { type: string }
        - in: query
          name: class
          schema: { type: string, enum: [info, warning, error] }
        - in: query
          name: source
          schema: { type: string }
        - in: query
          name: client
          schema: { type: string }
        - in: query
          name: open
          schema: { type: boolean }
        - in: query
          name: message_regex
          schema: { type: string }
      responses:
        '200': { description: Filtered events, stacked buckets, facets, type counts, and open-alert count }
  /home-stats:
    get:
      operationId: getHomeStats
      tags: [Home Alerts]
      summary: Get home-event and delivery counts
      parameters:
        - in: query
          name: from
          schema: { type: integer, format: int64, description: Unix seconds }
        - in: query
          name: to
          schema: { type: integer, format: int64, description: Unix seconds }
      responses:
        '200': { description: Counts for the selected period }
...
  /home-plugins/{kind}:
    put:
      operationId: configureHomePlugin
      tags: [Home Alerts]
      summary: Configure Telegram or Pushover (admin)
      parameters:
        - in: path
          name: kind
          required: true
          schema: { type: string, enum: [telegram, pushover] }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [destination, enabled]
              properties:
                destination: { type: string }
                secret: { type: string, description: Empty string keeps the existing token }
                enabled: { type: boolean }
      responses:
        '200': { description: Plugin configured; secret redacted }
    delete:
      operationId: deleteHomePlugin
      tags: [Home Alerts]
      summary: Remove plugin configuration (admin)
      parameters:
        - in: path
          name: kind
          required: true
          schema: { type: string, enum: [telegram, pushover] }
      responses:
        '204': { description: Plugin removed }
  /home-plugins/{kind}/test:
    post:
      operationId: testHomePlugin
      tags: [Home Alerts]
      summary: Send a test notification (admin)
      parameters:
        - in: path
          name: kind
          required: true
          schema: { type: string, enum: [telegram, pushover] }
      responses:
        '200': { description: Test message accepted by provider }
        '502': { description: Provider rejected the test message }
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: pagefire_session
      description: Server-side session cookie set by POST /auth/login
    bearerAuth:
      type: http
      scheme: bearer
      description: API token created via POST /auth/tokens
...
  responses:
    BadRequest:
      description: Bad request
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
...
  schemas:
    Error:
      type: object
      properties:
        error:
          type: string
      required: [error]
...
    User:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        email:
          type: string
          format: email
        role:
          type: string
          enum: [admin, user]
        timezone:
          type: string
          example: America/New_York
        avatar_url:
          type: string
        auth_provider:
          type: string
          nullable: true
        is_active:
          type: boolean
        last_login:
          type: string
          format: date-time
          nullable: true
        created_at:
          type: string
          format: date-time
      required: [id, name, email, role, timezone, is_active, created_at]
    CreateUserRequest:
      type: object
      required: [name, email]
      properties:
        name:
          type: string
        email:
          type: string
          format: email
        timezone:
          type: string
          default: UTC
        role:
          type: string
          enum: [admin, user]
          default: user
        password:
          type: string
          format: password
          description: If omitted, an invite token is generated instead.
    CreateUserResponse:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        email:
          type: string
        role:
          type: string
        timezone:
          type: string
        invite_url:
          type: string
          description: Present only when no password was provided.
...
    Team:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        description:
          type: string
        created_at:
          type: string
          format: date-time
      required: [id, name, created_at]
    CreateTeamRequest:
      type: object
      required: [name]
      properties:
        name:
          type: string
        description:
          type: string
...
    Service:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        description:
          type: string
        escalation_policy_id:
          type: string
        team_id:
          type: string
          nullable: true
        created_at:
          type: string
          format: date-time
      required: [id, name, escalation_policy_id, created_at]
    CreateServiceRequest:
      type: object
      required: [name, escalation_policy_id]
      properties:
        name:
          type: string
          maxLength: 200
        description:
          type: string
          maxLength: 2000
        escalation_policy_id:
          type: string
        team_id:
          type: string

    # ── Integration Keys ──
...
    EscalationPolicy:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        description:
          type: string
        repeat:
          type: integer
          minimum: 0
          maximum: 5
        team_id:
          type: string
          nullable: true
        created_at:
          type: string
          format: date-time
      required: [id, name, repeat, created_at]
    CreateEscalationPolicyRequest:
      type: object
      required: [name]
      properties:
        name:
          type: string
        description:
          type: string
        repeat:
          type: integer
          minimum: 0
          maximum: 5
          default: 0
        team_id:
          type: string
...
    Schedule:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        description:
          type: string
        timezone:
          type: string
        team_id:
          type: string
          nullable: true
        created_at:
          type: string
          format: date-time
      required: [id, name, timezone, created_at]
    CreateScheduleRequest:
      type: object
      required: [name]
      properties:
        name:
          type: string
        description:
          type: string
        timezone:
          type: string
          default: UTC
        team_id:
          type: string
...
    Alert:
      type: object
      properties:
        id:
          type: string
        service_id:
          type: string
        status:
          type: string
          enum: [triggered, acknowledged, resolved]
        summary:
          type: string
          maxLength: 500
        details:
          type: string
          maxLength: 10000
        source:
          type: string
          enum: [api, integration, monitor, test]
        dedup_key:
          type: string
        group_key:
          type: string
        escalation_policy_snapshot:
          type: string
          description: JSON-encoded snapshot of the escalation policy at alert creation time
        escalation_step:
          type: integer
        loop_count:
          type: integer
        next_escalation_at:
          type: string
          format: date-time
          nullable: true
        acknowledged_by:
          type: string
          nullable: true
        acknowledged_at:
          type: string
          format: date-time
          nullable: true
        resolved_at:
          type: string
          format: date-time
          nullable: true
        created_at:
          type: string
          format: date-time
      required: [id, service_id, status, summary, source, escalation_step, loop_count, created_at]
    CreateAlertRequest:
      type: object
      required: [service_id, summary]
      properties:
        service_id:
          type: string
        summary:
          type: string
          maxLength: 500
        details:
          type: string
          maxLength: 10000
        dedup_key:
          type: string
        group_key:
          type: string
...
    Incident:
      type: object
      properties:
        id:
          type: string
        title:
          type: string
          maxLength: 500
        status:
          type: string
          enum: [triggered, acknowledged, investigating, identified, monitoring, resolved]
        severity:
          type: string
          enum: [critical, major, minor]
        summary:
          type: string
          maxLength: 10000
        source:
          type: string
          enum: [manual, alert, monitor]
        created_by:
          type: string
        created_at:
          type: string
          format: date-time
        resolved_at:
          type: string
          format: date-time
          nullable: true
      required: [id, title, status, severity, source, created_at]
    CreateIncidentRequest:
      type: object
      required: [title]
      properties:
        title:
          type: string
          maxLength: 500
        status:
          type: string
          enum: [triggered, acknowledged, investigating, identified, monitoring, resolved]
          default: triggered
        severity:
          type: string
          enum: [critical, major, minor]
          default: critical
        summary:
          type: string
          maxLength: 10000
        source:
          type: string
          default: manual
...